Día 15/Proyecto personal textilBrain/Back-End-TextilBrain/maquila_backend/requirements.txt
asgiref==3.8.1
astroid==3.3.6
colorama==0.4.6
colored==2.2.4
dill==0.3.9
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
Faker==37.4.0
filelock==3.16.1
isort==5.13.2
joblib==1.5.1
mccabe==0.7.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.3.1
pandas==2.3.1
pillow==11.1.0
platformdirs==4.3.6
pylint==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tomlkit==0.13.2
tzdata==2025.1
virtualenv==20.29.1
